requirements.txt
amqp==5.0.9
asgiref==3.5.0
billiard==3.6.4.0
celery==5.2.3
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
Deprecated==1.2.13
dj-database-url==0.5.0
Django==4.0.1
django-cors-headers==3.11.0
django-filter==21.1
django-heroku==0.3.1
djangorestframework==3.13.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
kombu==5.2.3
numpy==1.22.2
packaging==21.3
pandas==1.4.0
prompt-toolkit==3.0.26
psycopg2==2.9.3
pyparsing==3.0.7
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.3
redis==4.1.2
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.31
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.8
vine==5.0.0
watchdog==2.1.6
wcwidth==0.2.5
whitenoise==5.3.0
wrapt==1.13.3
